remove warning when printing size_t
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@3425 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Kernel/src/simulationTools/EventDriven.cpp
@@ -120,9 +120,9 @@ void EventDriven::updateIndexSet(unsigned int i)
   assert(indexSet1);
   assert(indexSet2);
 
-  DEBUG_PRINTF("update indexSets start : indexSet0 size : %d\n", indexSet0->size());
-  DEBUG_PRINTF("update IndexSets start : indexSet1 size : %d\n", indexSet1->size());
-  DEBUG_PRINTF("update IndexSets start : indexSet2 size : %d\n", indexSet2->size());
+  DEBUG_PRINTF("update indexSets start : indexSet0 size : %d\n", (int)indexSet0->size());
+  DEBUG_PRINTF("update IndexSets start : indexSet1 size : %d\n", (int)indexSet1->size());
+  DEBUG_PRINTF("update IndexSets start : indexSet2 size : %d\n", (int)indexSet2->size());
 
   InteractionsGraph::VIterator uibegin, uipend, uip;
   boost::tie(uibegin, uipend) = indexSet0->vertices();
@@ -202,9 +202,9 @@ void EventDriven::updateIndexSet(unsigned int i)
     }
   }
 
-  DEBUG_PRINTF("update indexSets end : indexSet0 size : %d\n", indexSet0->size());
-  DEBUG_PRINTF("update IndexSets end : indexSet1 size : %d\n", indexSet1->size());
-  DEBUG_PRINTF("update IndexSets end : indexSet2 size : %d\n", indexSet2->size());
+  DEBUG_PRINTF("update indexSets end : indexSet0 size : %d\n", (int)indexSet0->size());
+  DEBUG_PRINTF("update IndexSets end : indexSet1 size : %d\n", (int)indexSet1->size());
+  DEBUG_PRINTF("update IndexSets end : indexSet2 size : %d\n", (int)indexSet2->size());
 }
 
 void EventDriven::updateIndexSetsWithDoubleCondition()
